requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
google-genai
pypdf2
python-pptx
python-dotenv
